Revert "refactor: mv routes' components to page-slice"
This reverts commit fcf58f7898a100b70efb201a703a9b2adc75e3d5.
