+++ src/routes/routes/add.tsx
@@ -14,12 +14,74 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { zodResolver } from '@hookform/resolvers/zod';
+import { notifications } from '@mantine/notifications';
+import { useMutation } from '@tanstack/react-query';
+import { createFileRoute, useRouter } from '@tanstack/react-router';
+import { FormProvider, useForm } from 'react-hook-form';
+import { useTranslation } from 'react-i18next';
 
-import { createFileRoute } from '@tanstack/react-router';
+import { postRouteReq } from '@/apis/routes';
+import { FormSubmitBtn } from '@/components/form/Btn';
+import { FormPartRoute } from '@/components/form-slice/FormPartRoute';
+import {
+  RoutePostSchema,
+  type RoutePostType,
+} from '@/components/form-slice/FormPartRoute/schema';
+import { FormTOCBox } from '@/components/form-slice/FormSection';
+import PageHeader from '@/components/page/PageHeader';
+import { req } from '@/config/req';
+import { produceRmUpstreamWhenHas } from '@/utils/form-producer';
+import { pipeProduce } from '@/utils/producer';
 
-import { RouteAdd } from '@/components/page-slice/routes/add';
+const RouteAddForm = () => {
+  const { t } = useTranslation();
+  const router = useRouter();
 
+  const postRoute = useMutation({
+    mutationFn: (d: RoutePostType) =>
+      postRouteReq(req, pipeProduce(produceRmUpstreamWhenHas('service_id'))(d)),
+    async onSuccess(res) {
+      notifications.show({
+        message: t('info.add.success', { name: t('routes.singular') }),
+        color: 'green',
+      });
+      await router.navigate({
+        to: '/routes/detail/$id',
+        params: { id: res.data.value.id },
+      });
+    },
+  });
+
+  const form = useForm({
+    resolver: zodResolver(RoutePostSchema),
+    shouldUnregister: true,
+    shouldFocusError: true,
+    mode: 'all',
+  });
+
+  return (
+    <FormProvider {...form}>
+      <form onSubmit={form.handleSubmit((d) => postRoute.mutateAsync(d))}>
+        <FormPartRoute />
+        <FormSubmitBtn>{t('form.btn.add')}</FormSubmitBtn>
+      </form>
+    </FormProvider>
+  );
+};
+
+function RouteComponent() {
+  const { t } = useTranslation();
+  return (
+    <>
+      <PageHeader title={t('info.add.title', { name: t('routes.singular') })} />
+      <FormTOCBox>
+        <RouteAddForm />
+      </FormTOCBox>
+    </>
+  );
+}
 
 export const Route = createFileRoute('/routes/add')({
-  component: RouteAdd,
+  component: RouteComponent,
 });
