+++ src/routes/routes/index.tsx
@@ -14,15 +14,120 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import type { ProColumns } from '@ant-design/pro-components';
+import { ProTable } from '@ant-design/pro-components';
 import { createFileRoute } from '@tanstack/react-router';
+import { useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 
-import { getRouteListQueryOptions } from '@/apis/hooks';
-import { RouteIndex } from '@/components/page-slice/routes/list';
+import { getRouteListQueryOptions, useRouteList } from '@/apis/hooks';
+import { DeleteResourceBtn } from '@/components/page/DeleteResourceBtn';
+import PageHeader from '@/components/page/PageHeader';
+import { ToAddPageBtn, ToDetailPageBtn } from '@/components/page/ToAddPageBtn';
+import { AntdConfigProvider } from '@/config/antdConfigProvider';
+import { API_ROUTES } from '@/config/constant';
 import { queryClient } from '@/config/global';
+import type { APISIXType } from '@/types/schema/apisix';
 import { pageSearchSchema } from '@/types/schema/pageSearch';
 
+const RouteList = () => {
+  const { data, isLoading, refetch, pagination } = useRouteList();
+  const { t } = useTranslation();
+
+  const columns = useMemo<ProColumns<APISIXType['RespRouteItem']>[]>(() => {
+    return [
+      {
+        dataIndex: ['value', 'id'],
+        title: 'ID',
+        key: 'id',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'name'],
+        title: t('form.basic.name'),
+        key: 'name',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'desc'],
+        title: t('form.basic.desc'),
+        key: 'desc',
+        valueType: 'text',
+      },
+      {
+        dataIndex: ['value', 'uri'],
+        title: 'URI',
+        key: 'uri',
+        valueType: 'text',
+      },
+      {
+        title: t('table.actions'),
+        valueType: 'option',
+        key: 'option',
+        width: 120,
+        render: (_, record) => [
+          <ToDetailPageBtn
+            key="detail"
+            to="/routes/detail/$id"
+            params={{ id: record.value.id }}
+          />,
+          <DeleteResourceBtn
+            key="delete"
+            name={t('routes.singular')}
+            target={record.value.id}
+            api={`${API_ROUTES}/${record.value.id}`}
+            onSuccess={refetch}
+          />,
+        ],
+      },
+    ];
+  }, [t, refetch]);
+
+  return (
+    <AntdConfigProvider>
+      <ProTable
+        columns={columns}
+        dataSource={data.list}
+        rowKey="id"
+        loading={isLoading}
+        search={false}
+        options={false}
+        pagination={pagination}
+        cardProps={{ bodyStyle: { padding: 0 } }}
+        toolbar={{
+          menu: {
+            type: 'inline',
+            items: [
+              {
+                key: 'add',
+                label: (
+                  <ToAddPageBtn
+                    key="add"
+                    label={t('info.add.title', { name: t('routes.singular') })}
+                    to="/routes/add"
+                  />
+                ),
+              },
+            ],
+          },
+        }}
+      />
+    </AntdConfigProvider>
+  );
+};
+
+function RouteComponent() {
+  const { t } = useTranslation();
+  return (
+    <>
+      <PageHeader title={t('sources.routes')} />
+      <RouteList />
+    </>
+  );
+}
+
 export const Route = createFileRoute('/routes/')({
-  component: RouteIndex,
+  component: RouteComponent,
   validateSearch: pageSearchSchema,
   loaderDeps: ({ search }) => search,
   loader: ({ deps }) =>
